requirements/requirements.txt
absl-py==2.1.0
aioredis==2.0.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.5
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
fastapi==0.115.0
fastapi-limiter==0.1.6
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
idna==3.8
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prompt_toolkit==3.0.50
protobuf==4.25.4
psycopg2==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyenv-win==3.1.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.1.1
requests==2.32.3
rich==13.8.0
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.24.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.38.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
